requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# MongoDB driver
motor>=3.3.2

# LangChain and AI integrations
langchain>=0.3.0
langchain-core>=0.3.7
langchain-google-genai>=2.0.1
langchain-openai>=0.2.2

# HTTP client for AI services
httpx>=0.25.2

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Production dependencies
gunicorn>=21.2.0
